Stack
Input is a string in the format "5 + 10 + POP 5 DUP -"
Create a C++ Program, where:
if any digit comes, it is pushed to stack
POP will result in removing top of the stack
DUP will duplicate the top of the stack
+ or - will result in popping two values from the top, calculating the
result and pushing the result at the top of the stack.
Final output should be the result after all operations described by the
string.